+++ ChangeLog.md
@@ -2,6 +2,18 @@
 
 **css** and **javascript** for a personnal website
 
+## [1.1.0] - In progress
+
+### Added
+
+- Add a Preact component to build the news page
+
+### Fixed
+
+- Refactoring sass style
+
+Initial version
+
 ## [1.0.0] - 2025-02-17
 
 - Refactoring sass style